# Fix API performance: one DNSSECKeeper per request