@zeha 2017-01-27
As said elsewhere, I think the `dnssec` flag can just go away, and with it goes the need for DNSSECKeeper in listZones.