@wojas 2017-01-27
@zeha The flag is currently being used. Removing it and requiring applications that need it for whatever reason to do individual calls for every single zone would not improve overal performance of the API.

How about adding an `?extra=...` GET parameter to this API call and only returning the `dnssec` field if you pass `?extra=dnssec`? This mechanism can also be easily extended to add other fields that the calling application might need (`?extra=dnssec,foo,bar`). The call takes slightly longer when the extra info is needed, but can save many individual requests that would otherwise be required. Users that don't need the extra fields simply do not provide it.

Such a change would be breaking for applications that use this field, but they can easily be fixed by adding the `extra` param. Adding this param to an application would also be backward compatible, as auth 4.0 will simply ignore it and always return the field.