@wojas 2017-01-27
One way to keep it backward compatible is to default to`?extra=dnssec` if the extra param is not set. To to a light query you could pass an empty `?extra=`.

Based on my personal quick testing removing the field improves performance by 10-15% after this patch. This is significant enough to provide a way to not incur the overhead, but not to completely remove the functionality.