@wojas 2017-01-31
@zeha I made the suggested change to pass UeberBackend to DNSSECKeeper, thanks!